Use `Integer` for indexes in `ClientBulkWriteResult.Verbose`. Make `ClientInsertOneResult.getInsertedId` null-able.
JAVA-5527

+++ driver-core/src/main/com/mongodb/client/result/bulk/ClientBulkWriteResult.java
@@ -110,7 +110,7 @@ public interface ClientBulkWriteResult {
          * @return The indexed {@link ClientInsertOneResult}s.
          * @see ClientBulkWriteException#getWriteErrors()
          */
-        Map<Long, ClientInsertOneResult> getInsertResults();
+        Map<Integer, ClientInsertOneResult> getInsertResults();
 
         /**
          * The indexed {@link ClientUpdateResult}s.
@@ -122,7 +122,7 @@ public interface ClientBulkWriteResult {
          * @return The indexed {@link ClientUpdateResult}s.
          * @see ClientBulkWriteException#getWriteErrors()
          */
-        Map<Long, ClientUpdateResult> getUpdateResults();
+        Map<Integer, ClientUpdateResult> getUpdateResults();
 
         /**
          * The indexed {@link ClientDeleteResult}s.
@@ -134,6 +134,6 @@ public interface ClientBulkWriteResult {
          * @return The indexed {@link ClientDeleteResult}s.
          * @see ClientBulkWriteException#getWriteErrors()
          */
-        Map<Long, ClientDeleteResult> getDeleteResults();
+        Map<Integer, ClientDeleteResult> getDeleteResults();
     }
 }
